Fix and update some trouble things.

+++ PudgeManga_Project/Data/migrations/20231211131442_Add_Rating_Model.cs
@@ -19,7 +19,7 @@ namespace PudgeManga_Project.Data.migrations
                     MangaId = table.Column<int>(type: "int", nullable: false),
                     AnimeId = table.Column<int>(type: "int", nullable: false),
                     UserId = table.Column<string>(type: "nvarchar(450)", nullable: false),
-                    Value = table.Column<int>(type: "int", nullable: false)
+                    Value = table.Column<double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
