+++ PudgeManga_Project/Data/migrations/ApplicationDBContextModelSnapshot.cs
@@ -482,8 +482,8 @@ namespace PudgeManga_Project.Data.migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(450)");
 
-                    b.Property<int>("Value")
-                        .HasColumnType("int");
+                    b.Property<double>("Value")
+                        .HasColumnType("float");
 
                     b.HasKey("Id");
